+++ doc/latex/compile.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-latexmk -pdf -output-directory="../"
+latexmk -pdf -output-directory="../" main.tex
 latexmk -c -output-directory="../"