+++ src/hamt/data.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../mesh/mesh2d_regular.h"
+#include "../mesh/mesh2d_triangular.h"
 
 namespace hamt {
 using namespace Eigen;
@@ -12,6 +13,7 @@ class Data {
     void Clear();
 
     Mesh2DRegular mesh2d_regular_;
+    Mesh2DTriangular mesh2d_triangular_;
     VectorXd results_;
 };
 }  // namespace hamt