+++ src/io/vtk/unstructured_grid.cc
@@ -106,6 +106,84 @@ void WritePointData(std::ofstream &stream, const std::vector<std::pair<std::stri
     stream << "</PointData>" << std::endl;
 }
 
+void WriteMesh2DTriangular(const Mesh2DTriangular &mesh, const std::string &file_name,
+                           const std::vector<std::pair<std::string, std::vector<double>>> &point_data) {
+    std::ofstream stream(file_name);
+
+    if (!stream.is_open()) {
+        throw Exception("Could not open file [" + file_name + "].", __PRETTY_FUNCTION__);
+    } else {
+        WriteHeader(stream);
+        WriteBody(stream, mesh, point_data);
+        WriteFooter(stream);
+    }
+
+    stream.close();
+}
+
+void WriteBody(std::ofstream &stream, const Mesh2DTriangular &mesh,
+               const std::vector<std::pair<std::string, std::vector<double>>> &point_data) {
+    stream << "<UnstructuredGrid>" << std::endl;
+    stream << "<Piece NumberOfPoints=\"" << mesh.nodes_.size() << "\" NumberOfCells=\"" << mesh.cells_.size() << "\">"
+           << std::endl;
+
+    WritePoints(stream, mesh);
+    WriteCells(stream, mesh);
+
+    if (point_data.size()) {
+        WritePointData(stream, point_data);
+    }
+
+    stream << "</Piece>" << std::endl;
+    stream << "</UnstructuredGrid>" << std::endl;
+}
+
+void WritePoints(std::ofstream &stream, const Mesh2DTriangular &mesh) {
+    stream << "<Points>" << std::endl;
+    stream << "<DataArray type=\"Float32\" NumberOfComponents=\"3\" format=\"ascii\">" << std::endl;
+
+    for (uint i = 0; i < mesh.nodes_.size(); i++) {
+        stream << mesh.nodes_.at(i).position(0) << " ";
+        stream << mesh.nodes_.at(i).position(1) << " ";
+        stream << mesh.nodes_.at(i).position(2) << " ";
+    }
+
+    stream << "</DataArray>" << std::endl;
+    stream << "</Points>" << std::endl;
+}
+
+void WriteCells(std::ofstream &stream, const Mesh2DTriangular &mesh) {
+    stream << "<Cells>" << std::endl;
+
+    stream << "<DataArray type=\"Int32\" Name=\"connectivity\" format=\"ascii\">" << std::endl;
+
+    for (uint i = 0; i < mesh.cells_.size(); i++) {
+        stream << mesh.cells_.at(i).nodes.at(0) << " ";
+        stream << mesh.cells_.at(i).nodes.at(1) << " ";
+        stream << mesh.cells_.at(i).nodes.at(2) << " ";
+    }
+
+    stream << "</DataArray>" << std::endl;
+
+    stream << "<DataArray type=\"Int32\" Name=\"offsets\" format=\"ascii\">" << std::endl;
+
+    for (uint i = 0; i < mesh.cells_.size(); i++) {
+        stream << (i + 1) * 3 << " ";
+    }
+
+    stream << "</DataArray>" << std::endl;
+
+    stream << "<DataArray type=\"UInt8\" Name=\"types\" format=\"ascii\">" << std::endl;
+
+    for (uint i = 0; i < mesh.cells_.size(); i++) {
+        stream << 5 << " ";
+    }
+
+    stream << "</DataArray>" << std::endl;
+
+    stream << "</Cells>" << std::endl;
+}
+
 }  // namespace unstructured_grid
 }  // namespace vtk
 }  // namespace hamt