+++ src/parser/opr_info/operation_info.cc
@@ -30,6 +30,11 @@ void OperationInfo::Execute(const std::vector<std::string>& argv, const bool&) {
                               << " ";
                     break;
                 }
+                case Mesh2DRegular::RADIATION: {
+                    std::cout << "TYPE: RADIATION"
+                              << " ";
+                    break;
+                }
             }
 
             std::cout << "VALUE: " << boundary.value << std::endl;
@@ -38,6 +43,46 @@ void OperationInfo::Execute(const std::vector<std::string>& argv, const bool&) {
         std::cout << "SURFACES: " << hamt_->data_->mesh2d_regular_.surfaces_.size() << std::endl;
         for (const auto& bin : hamt_->data_->mesh2d_regular_.surface_names_) {
             const Mesh2DRegular::Surface surface(hamt_->data_->mesh2d_regular_.surfaces_.at(bin.second));
+
+            std::cout << tab << bin.first << " ";
+            std::cout << "THERM. CONDUCT.: " << surface.thermal_conductivity << std::endl;
+        }
+    } else if (argv.at(0) == "triangl_mesh") {
+        std::cout << "TRIANGULAR MESH INFO" << std::endl;
+        std::cout << "NUMBER NODES : " << hamt_->data_->mesh2d_triangular_.nodes_.size() << std::endl;
+        std::cout << "NUMBER CELLS : " << hamt_->data_->mesh2d_triangular_.cells_.size() << std::endl;
+
+        std::cout << "BOUNDARIES: " << hamt_->data_->mesh2d_triangular_.boundaries_.size() << std::endl;
+        for (const auto& boundary_name_bin : hamt_->data_->mesh2d_triangular_.boundary_names_) {
+            const Mesh2DTriangular::Boundary boundary(
+                hamt_->data_->mesh2d_triangular_.boundaries_.at(boundary_name_bin.second));
+
+            std::cout << tab << boundary_name_bin.first << " ";
+
+            switch (boundary.type) {
+                case Mesh2DTriangular::DIRICHLET: {
+                    std::cout << "TYPE: DIRICHLET"
+                              << " ";
+                    break;
+                }
+                case Mesh2DTriangular::NEUMANN: {
+                    std::cout << "TYPE: NEUMANN"
+                              << " ";
+                    break;
+                }
+                case Mesh2DTriangular::RADIATION: {
+                    std::cout << "TYPE: RADIATION"
+                              << " ";
+                    break;
+                }
+            }
+
+            std::cout << "VALUE: " << boundary.value << std::endl;
+        }
+
+        std::cout << "SURFACES: " << hamt_->data_->mesh2d_triangular_.surfaces_.size() << std::endl;
+        for (const auto& bin : hamt_->data_->mesh2d_triangular_.surface_names_) {
+            const Mesh2DTriangular::Surface surface(hamt_->data_->mesh2d_triangular_.surfaces_.at(bin.second));
 
             std::cout << tab << bin.first << " ";
             std::cout << "THERM. CONDUCT.: " << surface.thermal_conductivity << std::endl;