+++ src/parser/opr_set/operation_set.cc
@@ -4,10 +4,12 @@ namespace hamt {
 
 OperationSet::OperationSet() : Operation("set") {
     std::shared_ptr<class SetRegularMesh> set_reg_mesh(std::make_shared<class SetRegularMesh>(name_));
+    std::shared_ptr<class SetTriangularMesh> set_triangl_mesh(std::make_shared<class SetTriangularMesh>(name_));
     std::shared_ptr<class SetWriter> set_writer(std::make_shared<class SetWriter>(name_));
     std::shared_ptr<class SetSolver> set_solver(std::make_shared<class SetSolver>(name_));
 
     sub_operations_[set_reg_mesh->GetName()] = set_reg_mesh;
+    sub_operations_[set_triangl_mesh->GetName()] = set_triangl_mesh;
     sub_operations_[set_writer->GetName()] = set_writer;
     sub_operations_[set_solver->GetName()] = set_solver;
 }