+++ src/parser/opr_set/operation_set.h
@@ -8,6 +8,7 @@
 #include "set_reg_mesh.h"
 #include "set_solver.h"
 #include "set_writer.h"
+#include "set_trinagl_mesh.h"
 
 namespace hamt {
 class OperationSet : public Operation {