+++ src/solver/solver.cc
@@ -17,6 +17,10 @@ void Solver::Execute() {
                     ExecuteHomogenRegMesh();
                     break;
                 }
+                case TRIANGULAR: {
+                    ExecuteHomogenTrianglMesh();
+                    break;
+                }
                 default: {
                     throw Exception("Undefined mesh type" + std::to_string(config_.mesh_type), __PRETTY_FUNCTION__);
                 }
@@ -48,6 +52,28 @@ void Solver::ExecuteHomogenRegMesh() {
             break;
         }
         default: { throw Exception("Undefined coord type" + std::to_string(config_.coord_type), __PRETTY_FUNCTION__); }
+    }
+}
+
+void Solver::ExecuteHomogenTrianglMesh() {
+    if (static_cast<size_t>(data_->results_.rows()) != data_->mesh2d_triangular_.nodes_.size()) {
+        data_->results_ = VectorXd::Zero(data_->mesh2d_triangular_.nodes_.size());
+    }
+
+    switch (config_.coord_type) {
+        case CARTESIAN: {
+            const std::pair<MatrixXd, VectorXd> mat_b(heat_equation_homogeneous::ConvertMesh2dTriangularCartesian(
+                data_->mesh2d_triangular_, data_->results_));
+            data_->results_ = mat_b.first.partialPivLu().solve(mat_b.second);
+            break;
+        }
+        case CYLINDER: {
+            throw IncompleteCodeError("cylinder coordinates not defined for trangular mesh");
+            break;
+        }
+        default: {
+            throw Exception("Undefined coord type" + std::to_string(config_.coord_type), __PRETTY_FUNCTION__);
+        }
     }
 }